- lint - add database_test skeleton for testing - add auth test type - extend account structure

+++ db/schema/auth.sql
@@ -1,6 +1,6 @@
 CREATE TABLE account(
     address TEXT UNIQUE PRIMARY KEY NOT NULL,
-    balance NUMERIC NOT NULL,
+    balance BIGINT NOT NULL,
     code TEXT NOT NULL,
     keys_list TEXT,
     contract_map TEXT
@@ -9,13 +9,13 @@ CREATE TABLE account(
 CREATE TABLE locked_account(
     account_address TEXT UNIQUE PRIMARY KEY NOT NULL REFERENCES account(address),
     locked_address TEXT NOT NULL,
-    balance NUMERIC NOT NULL,
-    unlock_limit NUMERIC NOT NULL
+    balance BIGINT NOT NULL,
+    unlock_limit BIGINT NOT NULL
 );
 
 CREATE TABLE delegator_account(
     account_address TEXT UNIQUE PRIMARY KEY NOT NULL REFERENCES account(address),
-	delegator_id    INTEGER NOT NULL,
+	delegator_id    BIGINT NOT NULL,
 	delegator_node_id   TEXT NOT NULL,
 	delegator_node_info TEXT NOT NULL
 );
